Rem-lapse/requirements.txt
discord.py
python-dotenv
better-profanity
PyNaCl
yt-dlp
aiohttp